desktop/requirements.txt
PyVISA==1.9.1
numpy==1.15.4
pyserial==3.4
